// src/app/api/generated/model/importar-aranceles-request.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FilaImportacionArancelRequest } from './fila-importacion-arancel-request';


/**
 * Lote de aranceles a previsualizar o confirmar
 */
export interface ImportarArancelesRequest { 
    /**
     * Las filas, en el orden de la planilla. La posicion (desde 1) es la que identifica cada fila en la respuesta
     */
    filas: Array<FilaImportacionArancelRequest>;
    /**
     * PREVIEW valida cada fila sin escribir ni bloquear. CONFIRMAR revalida todo bajo el lock del convenio y aplica el lote entero, o nada
     */
    modo: ImportarArancelesRequestModoEnum;
}
export enum ImportarArancelesRequestModoEnum {
    PREVIEW = 'PREVIEW',
    CONFIRMAR = 'CONFIRMAR'
};